requirements.txt
alabaster==0.7.12
anyio==3.7.0
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
exceptiongroup==1.1.1
fastapi==0.98.0
fasteners==0.14.1
feedparser==6.0.8
future==0.18.2
gitsome==0.8.0
h11==0.14.0
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpydoc==1.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.9
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.4.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
roman==3.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==4.3.2
starlette==0.27.0
systemd-python==234
texttable==1.6.4
typing_extensions==4.6.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
xonsh==0.11.0
zipp==1.0.0
